Fixed IFTPP.
1. Fixed three sources of data leakage:
(1) For the parameterized distribution, the running statistics of dts were computed using the current batch's dts which effectively cheats by knowing ahead of time the scale of the targeted dts.
(2) The shifting of dts misaligned with the marks.
(3) The log-likelihood of marks is computed against the wrong marks due to incorrect shifting - specifically the marks that they conditioned on.

2. Corrected the log likelihood that was reported on an incorrect scale. It originally was ll.sum(dim=-1).mean() when it should just be ll.sum().

3. Now supports next event prediction based on the parameterized distribution.

Co-Authored-By: ajboyd2 <11725441+ajboyd2@users.noreply.github.com>